Move cairo0 contract utils to integration crate
